requirements.txt
pandas
matplotlib
seaborn
statsmodels
requests
pandas
python-dotenv
